Go back to the original command K button
